+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Updated
 - updated generated app to simple & unopinionated react app w/ parcel
 - `yeoman-generator` to latest version w/ adopting changes.
+- `yarn` as default package manager w/ supported `--npm` flag.
 
 ## [1.7.1][] - 2021-02-03
 ### Fixed